pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pokemon-md-agent"
version = "0.1.0"
description = "AI agent for Pokemon Mystery Dungeon: Red Rescue Team"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Games/Entertainment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "asyncio-mqtt>=0.13.0",
    "websockets>=12.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "faiss-cpu>=1.7.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "unsloth",
    "unsloth-zoo",
    "bitsandbytes>=0.41.0",
    "qwen-vl-utils>=0.0.3",
    "sentence-transformers>=2.2.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "imagehash",
    "scikit-image>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]

[project.scripts]
pmd-agent = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-dir]
"pokemon_md_agent" = "src"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["S101", "S104", "COM812", "ISC001"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["PIL.*", "faiss.*", "sentence_transformers.*", "transformers.*", "torch.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -rfEsx --maxfail=1 --timeout=30 --timeout-method=thread -m 'not slow and not network and not bench and not longctx'"
filterwarnings = ["ignore::DeprecationWarning"]
markers = ["slow","network","bench","longctx"]